docs: Update pull request template
Update the pull request template to provide more comprehensive guidelines for submitting a pull request. The updated template now includes a checklist of tasks, instructions on linking to related issues, and a template for release notes. This will help ensure that all necessary information is provided when submitting a pull request.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>